separate inputs from fullstack type & update docs
also, route ui traffic to /*
